+++ src/modules/delay/apps/prompt.svelte
@@ -8,7 +8,7 @@
 				{@const isSelected=t.id === selected}
 				<div
 				  class={[
-						"flex items-center h-[50px] px-1 border",
+						"flex items-center h-[50px] pl-1 pr-2 border",
 						disabled ? "cursor-not-allowed rounded-none" : "cursor-pointer rounded-sm",
 						isSelf ? "dark:bg-green-300/20 bg-green-600/20 border-green-400" :
 						disabled ? "dark:bg-white/10 bg-black/10 border-transparent" :
@@ -16,9 +16,9 @@
 					]}
 					onclick={() => { if (!disabled) selected = t.id }}
 				>
-					<img class="h-full p-0.5
-					" {...imgPropsForToken(t.token!.object!)} inert>
-					<p class="grow overflow-hidden text-ellipsis max-w-[20ch] max-h-[48px] ml-0.5" style="display: -webkit-box; -webkit-line-clamp: 3; -webkit-box-orient: vertical;">{displayName(t)}</p>
+					<img class="h-full p-0.5" {...imgPropsForToken(t.token!.object!)} inert>
+					<p class="grow overflow-hidden text-ellipsis max-w-[20ch] max-h-[48px] ml-0.5" style="display: -webkit-box; -webkit-line-clamp: 2; -webkit-box-orient: vertical;">{displayName(t)}</p>
+					<p class="ml-3">{t.initiative}</p>
 				</div>
 				{:else}
 					<p class="text-center">Empty combat?</p>